pyproject.toml
[tool.poetry]
name = "film-search-engine"
version = "0.1.0"
description = "Natural-language movie search engine using sentence-transformers, FAISS, FastAPI, and Streamlit"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
	{ include = "src" },    # package the src/ namespace so imports like 'from src...' work
	{ include = "scripts" }, # package helper scripts (optional)
	{ include = "tests" }    # package tests (optional)
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
# Core data + math
pandas = ">=1.3.0"
numpy = ">=1.21.0"
# NLP / Embeddings
sentence-transformers = ">=2.2.0"
spacy = ">=3.5.0"
# Vector search
faiss-cpu = ">=1.7.4"
# API / Web
fastapi = ">=0.115.0"
uvicorn = {version = ">=0.30.0", extras = ["standard"]}
requests = ">=2.31.0"
streamlit = ">=1.36.0"
# Utilities
loguru = ">=0.7.2"
rapidfuzz = ">=2.0.0"
click = ">=8.0.0"
# TF runtime (indirect via transformers in some envs)
# Keep flexible; Poetry will resolve compatible version available for platform
# It can be removed if not needed in your environment
"tf-keras" = "*"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
black = ">=23.0.0"
isort = ">=5.12.0"

[tool.poetry.scripts]
# Convenience launchers (run with: poetry run movie-search-api)
movie-search-api = "uvicorn:main"

[tool.uvicorn]
# Helper for scripts: poetry run uvicorn api:app --reload
app = "api:app"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
